feat(avo-1912): add mobile layout for assignment metadat, fix pupil name
